+++ common/src/main/java/io/github/mortuusars/exposure_catalog/data/ExposureThumbnail.java
@@ -1,19 +1,28 @@
 package io.github.mortuusars.exposure_catalog.data;
 
+import io.github.mortuusars.exposure.render.image.IImage;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.world.level.material.MapColor;
 
 import java.util.Arrays;
 import java.util.Objects;
 
-public class ExposureThumbnail {
-    protected int width, height;
+public class ExposureThumbnail implements IImage {
+    private final String id;
+    protected int width;
+    protected int height;
     protected byte[] pixels;
 
-    public ExposureThumbnail(int width, int height, byte[] pixels) {
+    public ExposureThumbnail(String id, int width, int height, byte[] pixels) {
+        this.id = id;
         this.width = width;
         this.height = height;
         this.pixels = pixels;
+    }
+
+    @Override
+    public String getImageId() {
+        return id;
     }
 
     public int getWidth() {
@@ -26,10 +35,6 @@ public class ExposureThumbnail {
 
     public byte[] getPixels() {
         return pixels;
-    }
-
-    public boolean isEmpty() {
-        return width <= 0 || height <= 0 || pixels.length == 0;
     }
 
     public byte getPixel(int x, int y) {
@@ -56,12 +61,13 @@ public class ExposureThumbnail {
     }
 
     public void toBuffer(FriendlyByteBuf buffer) {
+        buffer.writeUtf(id);
         buffer.writeInt(width);
         buffer.writeInt(height);
         buffer.writeByteArray(pixels);
     }
 
     public static ExposureThumbnail fromBuffer(FriendlyByteBuf buffer) {
-        return new ExposureThumbnail(buffer.readInt(), buffer.readInt(), buffer.readByteArray());
+        return new ExposureThumbnail(buffer.readUtf(), buffer.readInt(), buffer.readInt(), buffer.readByteArray());
     }
 }