+++ common/src/main/java/io/github/mortuusars/exposure_catalog/network/handler/ClientPacketsHandler.java
@@ -33,7 +33,7 @@ public class ClientPacketsHandler {
 
     public static void receiveExposureThumbnail(SendExposureThumbnailS2CP packet) {
         executeOnMainThread(() -> {
-            CatalogClient.setThumbnail(packet.exposureId(), packet.thumbnail());
+            CatalogClient.setThumbnail(packet.thumbnail());
         });
     }
 }