+++ CHANGELOG.md
@@ -1,1 +1,5 @@
 # Changelog
+
+# 1.0.1 - Unreleased
+[Requires Exposure 1.7.3+]
+- Fixed error when trying to export exposures on a dedicated server.